+++ src/components/tagAnalyzer/OverlapModal.scss
@@ -26,17 +26,13 @@
             display: flex;
             align-items: center;
         }
-        padding: 0 20px;
+        padding: 0 15px 0 20px;
         font-size: 14px;
         width: 100%;
         height: 10%;
         display: flex;
         align-items: center;
         justify-content: space-between;
-        svg {
-            cursor: pointer;
-            margin-right: 4px;
-        }
     }
     .Overlap-body {
         width: 100%;
